// src/app/models/comments.ts
/*Comments*/ 
export interface IRegisterComments {
    comment: string;
    rating: number; 
    date: string;
    idBooking: string; 
    idUser: string;
  }
  
  export interface ICommentsEntity {
    comment: string;
    rating: number; 
    date: string;
    idBooking: string; 
    idUser: string;
  }
  